refactor: added link to chracter on fluentApi, separate ItemAccordion
